Implement modules and move common code to `internal` package
