+++ go.mod
@@ -1,0 +1,14 @@
+module mctester
+
+go 1.18
+
+require (
+	github.com/jamiealquiza/tachymeter v2.0.0+incompatible
+	go.uber.org/ratelimit v0.2.0
+	golang.org/x/sync v0.0.0-20220601150217-0de741cfad7f
+)
+
+require (
+	github.com/andres-erbsen/clock v0.0.0-20160526145045-9e14626cd129 // indirect
+	github.com/dgryski/go-pcgr v0.0.0-20211101192959-4b34ab9ccb8c // indirect
+)
